Code for Ants
Beginning of the addition of ants. Adds the code for ants, a new colony hostile mob that'll take the place of roaches as a swarm faction. They've got the exact same stats as roaches except they're a different mob type so we can edit them a bit better, and they're also missing a few of the annoying types. They provide all the same chems that roaches do too.