realization of the true potential of the guild mop (#182)
## About The Pull Request

Re-adds core functionality of the guild mop that was removed at some
point: giving it its condenser back.

The guild mop now takes a medium cell and can generate small quantities
of space cleaner, making it a faster alternative to the vacuum (that
doesn't generate biomass).

## Changelog
:cl:
balance: The guild mop now takes a medium cell and can generate small
quantities of space cleaner, making it a faster alternative to the
vacuum (that doesn't generate biomass).
/:cl: